Fix bad tail state in LRU map on removal of last entry
On removal of the last entry from an LRU map, the entry is both _head
and _tail. The existing remove() implementation matched it against
_head, which was then updated to _head.next (null), but _tail was left
unmodified in a bad state.

Since _tail is not null, the next insertion into the newly-empty map
don't reset it. When the map size eventually exceeds the maximum size,
_removeLru() is called to evict _tail, resulting in a null-pointer
exception as _removeLru() attempts to get _tail.previous (null) and then
set .next on that.