LruMap: Update the previous reference of the next entry when removing entries
When removing an entry in the middle of the linked list the previous pointer
in the following entry is not updated. When this pointer is used the LruMap
can reach a state where there are entries in the HashMap but not in the
LinkedList.

Any such entries will never be evicted from the cache since cache eviction
comes from the LinkedList.