Fix memory leak in LruCache (#389)
* LruMap promotion of the MRU entry causes the entry to link to itself

Fixes https://github.com/google/quiver-dart/issues/357

When the most-recently-used element in the map is re-added, promotion
caused a loop in the linked list. In such a case, the correct behaviour
is no-op.

* LruMap: Update the previous reference of the next entry when removing entries

When removing an entry in the middle of the linked list the previous pointer
in the following entry is not updated. When this pointer is used the LruMap
can reach a state where there are entries in the HashMap but not in the
LinkedList.

Any such entries will never be evicted from the cache since cache eviction
comes from the LinkedList.

* Add Adam Lofts to authors list.