Run dartdevc-compiled web tests on Travis (#448)
Re-enables DDC web tests on Travis.

The build_runner package makes DDC testing quick and easy, but
indirectly depends on quiver, which introduces a circular dependency. To
avoid this, we patch the pubspec for the test run only.

This change also sets `env` to a single variable rather that a list of one var.