Narrow generate_for glob in build.yaml (#498)
The existing generate_for was overriding the default glob to also include the `lib` directory, which was slowing down initial builds a tiny bit for all users who have a transitive dep on this package. (it wouldn't actually compile them since they aren't apps, but it was running on all dart files in lib and skipping them).

This removes build.yaml entirely, updating build_web_compilers.