Instrument pending timers in tests
Flutter widget tests assert if a test completes with timers still
pending.  However, it can be hard to diagnose where a pending timer
came from.  For example, a widget might consume a third-party library
that internally uses a timer.

To facilitate this, modify quiver's FakeAsync to store
StackTrace.current when constructing each _FakeTimer.  Add a
pendingTimersDebugInfo getter to to report this.

Fixes #421.