Use isEmpty rather than checks vs 0 where possible
This also adds TreeSet.isEmpty and TreeSet.isNotEmpty in order to fix a
stack overflow error when TreeSet.isEmpty or TreeSet.isNotEmpty are
called. Since TreeSet extends IterableBase, the default implementation
of isEmpty is to obtain an iterator over itself and check the return
value of iterable.moveNext().

Enables the prefer_is_empty lint.