Eliminate classImplements TODO
This suggested implementation cleanup when htts://dartbug.com/19781 was
fixed. The bug in question was filed on the assumption that a == b in the
following example:

   class Foo implements Comparable {}

   var a = reflect(Foo).superinterfaces
               .singleWhere((e) => e.simpleName == #Comparable);
   var b = reflectClass(Comparable)

However, in this example a is a mirror on Comparable<dynamic> (an
instantiation of a generic class declaration) whereas b is a mirror on
Comparable (a generic class declaration).