Use isEmpty rather than checks vs 0 where possible (#563)
This also adds TreeSet.isEmpty and TreeSet.isNotEmpty in order to fix a
stack overflow error when TreeSet.isEmpty or TreeSet.isNotEmpty are
called in the iterator's moveNext implementation.

Since TreeSet extends IterableBase, and didn't override isEmpty, it fell back
on the IterableBase implementation, which is to obtain an iterable over itself
and check the return value of calling moveNext. Since moveNext now calls
isEmpty, we drop into an infinite recurive loop. A quick length check
implementation is a significant performance improvement as well.

Enables the prefer_is_empty lint.