Eliminate bad use of const in Optional.transform (#587)
It's invalid to create a const which references a type variable. The
language team decided that in order to avoid painful migration of common
Dart 1 usages such as `const []`, rather than generate an error, they
would replace the type variable with `Null`, since usually, for properly
covariant types, the result 'works'.

In cases where `transform` is chained with e.g., `or`, you can end up
with `Optional<T>.absent` replacing `T` with `Null`, but then do an
implicit covariance check on the incoming type in `or`, which should be
the `S` in `transform` but is found to be `Null`.

As a result, we revert the move to const construction in `transform` and
`transformNullable`.

This reverts this bit of constification that was added in e769045.