Fix unnecessary use of null as initial value
When testing Metronome, we maintain a lastTime variable which we update
on each beat of the metronome, then later test against expected updates.
We were previously initializing this to null, but it's equally
reasonable to set it to the clock start time as the time at 'zero-th'
(minus-first?) beat and avoid the null.