Add null annotations to core.Optional
This adds nullability assertions to Optional in preparation for
migration to non-null by default in Dart.

The reason that we extend IterableBase<T> rather than IterableBase<T?>
is that Optional.expand eliminates any nulls.

Once NNBD is available, the utility of this class is limited and we
should consider deprecating it.