Move stringFromByteStream to async; deprecate IO (#628)
This deprecates two of the three function in the IO library and migrates
the remaining function (which doesn't use dart:io) to the async library.

As discussed in https://github.com/google/quiver-dart/issues/621, a
survey of ~150,000 imports of quiver at Google suggests that only
stringFromByteStream is used internally. It has been moved to the async
library.

getFullPath can be replaced with File.resolveSymbolicLinksSync if used.

visitDirectory should be rewritten if any users are making use of it.
Alternatively, it and its tests can be copied to other codebases under
the terms of the Apache 2.0 license.

Eliminating these functions avoids pub.dev suggesting that quiver can't
be used for web projects.