Allow null args to lists/maps/setsEqual
This updates the parameters to listsEqual, mapsEqual and setsEqual in
the collection utils library to nullable types.

This avoids forcing callers to manually check for null, is a bit more in
line with Object.equals() which handles the null check in the runtime,
and avoids a breaking change to users who have not yet transitioned to
non-null by default.